back_end/requirements.txt
fastapi
uvicorn
supabase
openai
python-dotenv
pydantic
python-multipart
cors
